Move `formatError` to `browser-logs`
