+++ packages/browser-logs/package.json
@@ -47,10 +47,13 @@
   ],
   "dependencies": {
     "@web/test-runner": "^0.15.0",
-    "errorstacks": "^2.2.0"
+    "diff": "^5.0.0",
+    "errorstacks": "^2.2.0",
+    "nanocolors": "^0.2.1"
   },
   "devDependencies": {
     "@esm-bundle/chai": "^4.1.5",
+    "@types/diff": "^5.0.0",
     "puppeteer": "^13.1.2"
   }
 }
