+++ packages/browser-logs/src/index.ts
@@ -1,3 +1,4 @@
 export { deserialize } from './deserialize';
 export { browserScript } from './browserScript';
 export { parseStackTrace, StackLocation, MapStackLocation } from './parseStackTrace';
+export { formatError } from './format-error';
