+++ packages/test-runner/src/index.ts
@@ -7,5 +7,4 @@ export { startTestRunner } from './startTestRunner';
 export { defaultReporter } from './reporter/defaultReporter';
 export { summaryReporter } from './reporter/summaryReporter';
 export { dotReporter } from './reporter/dotReporter';
-export { formatError } from './reporter/reportTestsErrors';
 export type TestRunnerConfig = Partial<FullTestRunnerConfig>;
